Refactor settings menu to use SwitchSetting subclasses and factory
