+++ src/config_menu/switches/tcp_reno.ts
@@ -1,0 +1,11 @@
+import { GlobalContext } from "../../context";
+import { SwitchSetting } from "./switch";
+
+export class UseTcpRenoSwitch extends SwitchSetting {
+  constructor(ctx: GlobalContext) {
+    super(ctx, "useTcpReno", "Use TCP Reno", true);
+  }
+  apply() {
+    this.ctx.setUseTcpReno(this.value);
+  }
+}
